Fix early init of MapperFactoryBean
This commit post-processes the bean factory so that the target type
produced by the factory bean is exposed. While this is not a problem
with a regular runtime, AOT uses instance supplier and without a proper
bound for the generic type that it produces, the only solution is to
create it early.